STY: Change Link to be a non-markup annotation Section 12.5.6 of the PDF 2.0 specification
